+++ backend/src/main/java/com/studysync/creational/factorymethod/CreditCardFactory.java
@@ -1,8 +1,0 @@
-package com.studysync.creational.factorymethod;
-
-public class CreditCardFactory extends PaymentProcessorFactory {
-    @Override
-    public PaymentProcessor createProcessor() {
-        return new CreditCardProcessor();
-    }
-}