+++ backend/src/main/java/com/studysync/creational/factorymethod/CryptoFactory.java
@@ -1,8 +1,0 @@
-package com.studysync.creational.factorymethod;
-
-public class CryptoFactory extends PaymentProcessorFactory {
-    @Override
-    public PaymentProcessor createProcessor() {
-        return new CryptoProcessor();
-    }
-}