+++ backend/src/main/java/com/studysync/creational/factorymethod/PayPalFactory.java
@@ -1,8 +1,0 @@
-package com.studysync.creational.factorymethod;
-
-public class PayPalFactory extends PaymentProcessorFactory {
-    @Override
-    public PaymentProcessor createProcessor() {
-        return new PayPalProcessor();
-    }
-}